Separate trie from the rest of the storage code
