+++ core/storage/src/test_utils.rs
@@ -2,7 +2,8 @@ use std::sync::Arc;
 
 use kvdb_memorydb::InMemory;
 
-use crate::{StateDb, TOTAL_COLUMNS};
+use crate::state_db::StateDb;
+use crate::TOTAL_COLUMNS;
 
 pub type MemoryStorage = InMemory;
